Move dialogs from widgets to views package.
Some dialogs were in one, some in the other...
