+++ gapic/src/main/com/google/gapid/views/GotoMemory.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.gapid.widgets;
+package com.google.gapid.views;
 
 import static com.google.gapid.util.Paths.memoryAfter;
 import static com.google.gapid.widgets.Widgets.createComposite;
